pledge.html
I changed the initial color of the map and added gradations of color in a legend. These color gradations will correspond to the number of pledges with the color of each county changing (when hovered over with a mouse) based upon participation once the data from the form is incremented and  linked to the map script.